+++ CHANGELOG.md
@@ -29,6 +29,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [CVE-2026-53404](https://nvd.nist.gov/vuln/detail/CVE-2026-53404) and
   [CVE-2026-50229](https://nvd.nist.gov/vuln/detail/CVE-2026-50229).
 
+### Added
+- [#356] Add API and persistence for personal access tokens.
+
 ## [v7.3.7-4] - 2026-07-24
 ### Fixed
 - [#352] prevent ``config/_global/certificate/additional/toc`` from having multiple identical keys by throwing an error